django-api/requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
attrs==21.2.0
cattrs==1.7.0
Django==3.2.3
django-api==0.1.3
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
Faker==8.5.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
iniconfig==1.1.1
mixer==7.1.2
packaging==20.9
pluggy==0.13.1
promise==2.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
python-dateutil==2.8.1
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
